requirements.txt
flask
sshtunnel
mysql-connector-python
python-dotenv
pandas
pyodbc
sqlalchemy
